+++ src/pages/home.astro
@@ -1,5 +1,5 @@
 ---
-import HomePage from '../components/home';
+import HomePage from '../components/HomePage';
 import Layout from '../layouts/Layout.astro';
 ---
