+++ src/pages/login.astro
@@ -1,5 +1,5 @@
 ---
-import LoginPage from '../components/login';
+import LoginPage from '../components/LoginPage';
 import Layout from '../layouts/Layout.astro';
 ---
