+++ src/pages/signup.astro
@@ -1,5 +1,5 @@
 ---
-import SignupPage from '../components/signup';
+import SignupPage from '../components/SignupPage';
 import Layout from '../layouts/Layout.astro';
 ---
